Update IOC protocol file
Add the current version of the .proto file being used in the IOC.
This updated protocol file matches the more informative header
(`[command][status][data]\r\n`) discussed in previous commits.
Also, this new header provides a mismatch handler that is
responsible for detecting and handling out-of-sync messages,
ensuring proper synchronization between the IOC and BeagleBone.